+++ tests/test_flysight_v1_parser.py
@@ -20,7 +20,7 @@ def test_log_parse() -> None:
 
     DERIVED_COLS = ("time", "elapsed_time", "groundspeed")
     for col_name in DERIVED_COLS:
-        checks.is_col(flight_log, col_name)
+        checks.is_col(flight_log.track_data, col_name)
 
     truth_dt = dt.datetime.strptime("2021-04-20T12:34:20.00Z", r"%Y-%m-%dT%H:%M:%S.%f%z")
     TRUTH_DERIVED = polars.DataFrame(
@@ -30,7 +30,25 @@ def test_log_parse() -> None:
             "groundspeed": [1.165],
         }
     )
-    assert_frame_equal(flight_log.select(DERIVED_COLS), TRUTH_DERIVED, rtol=1e-3)
+    assert_frame_equal(flight_log.track_data.select(DERIVED_COLS), TRUTH_DERIVED, rtol=1e-3)
+
+
+def test_device_info_parse() -> None:
+    sample_flight_log = SAMPLE_DATA_DIR / "21-04-20.CSV"  # Test log with single data row
+    flight_log = flysight_proc.load_flysight(sample_flight_log)
+
+    assert flight_log.device_info.flysight_type == FlysightType.VERSION_1
+
+    # fmt: off
+    truth_sensor_info = flysight_proc.SensorInfo(
+        columns=["time", "lat", "lon", "hMSL", "velN", "velE", "velD", "hAcc", "vAcc", "sAcc", "heading", "cAcc", "gpsFix", "numSV"],
+        units=["datetime", "deg", "deg", "m", "m/s", "m/s", "m/s", "m", "m", "m/s", "deg", "deg", "", ""],
+        id_="GNSS",
+    )
+    # fmt: on
+
+    assert "GNSS" in flight_log.device_info.sensor_info.keys()
+    assert flight_log.device_info.sensor_info["GNSS"] == truth_sensor_info
 
 
 def test_log_parse_normalize_gps() -> None:
@@ -38,8 +56,18 @@ def test_log_parse_normalize_gps() -> None:
     flight_log = flysight_proc.load_flysight(sample_flight_log, normalize_gps=True)
 
     # Normalization helper tested elsewhere, just check here that the flag is being acted on
-    assert flight_log["lat"][0] == pytest.approx(0)
-    assert flight_log["lon"][0] == pytest.approx(0)
+    assert flight_log.track_data["lat"][0] == pytest.approx(0)
+    assert flight_log.track_data["lon"][0] == pytest.approx(0)
+
+
+def test_log_parse_normalize_gps_method() -> None:
+    sample_flight_log = SAMPLE_DATA_DIR / "21-04-20.CSV"  # Test log with single data row
+    flight_log = flysight_proc.load_flysight(sample_flight_log, normalize_gps=False)
+    flight_log.normalize_gps()
+
+    # Normalization helper tested elsewhere, just check here that the flag is being acted on
+    assert flight_log.track_data["lat"][0] == pytest.approx(0)
+    assert flight_log.track_data["lon"][0] == pytest.approx(0)
 
 
 def test_batch_log_parse() -> None:
@@ -61,8 +89,8 @@ def test_batch_log_parse_normalize_gps() -> None:
 
     # Normalization helper tested elsewhere, just check here that the flag is being acted on
     for fl in flight_logs["sample_data"].values():
-        assert fl["lat"][0] == pytest.approx(0)
-        assert fl["lon"][0] == pytest.approx(0)
+        assert fl.track_data["lat"][0] == pytest.approx(0)
+        assert fl.track_data["lon"][0] == pytest.approx(0)
 
 
 SAMPLE_DATA_TO_SPLIT = dedent(