+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<major>`.`<minor>`.`<patch>`)
 
+## [0.7.0]
+### Added
+* #15 Add `pyflysight.flysight_proc.FlysightV1` and `pyflysight.flysight_proc.FlysightV1FlightLog` container classes for interfacing with FlySight V1 track data & metadata
+* #30 Add optional normalization of GPS coordinates in plaintext log files
+* #34 Add CLI pipeline for parsing FlySight V2 flight logs into more user-friendly CSV files.
+
+### Changed
+* (Internal) #21 Utilize MkDocs for documentation generation
+* #15 FlySight V1 related parsing helpers now utilize `pyflysight.flysight_proc.FlysightV1FlightLog` instances rather than bare `DataFrame`s
+* (Internal) Migrate to uv from poetry
+
 ## [0.6.0]
 ### Added
 * #24 Add optional normalization of parsed GPS coordinates
@@ -23,10 +34,11 @@ Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<ma
 
 ### Added
 * Add additional helpers to `pyflysight.config_utils` and `pyflysight.flysight_utils`
+* (Internal) #21 Add documentation autogeneration using `pdoc3`
 
 ## [v0.4.0]
 ### Added
-* Add derived `total_accel` column to Flysight V2 IMU sensor dataframe, calculated as a vector sum of the `xyz` acceleration components
+* Add derived `total_accel` column to FlySight V2 IMU sensor dataframe, calculated as a vector sum of the `xyz` acceleration components
 * Add `pyflysight.log_utils.locate_log_subdir` helper for resolving child log directory from a given top-level directory
 * Add `pyflysight.log_utils.iter_log_dirs` helper for iterating through child log directories of a given top-level directory
 * #19 Add `pyflysight.config_utils` for config file generation
@@ -37,14 +49,14 @@ Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<ma
 * (Internal) Bump to Polars v1.x
 
 ### Added
-* #18 Add `pyflysight.log_utils.classify_log_dir` helper for classifying the Flysight hardware rev of a given log directory
+* #18 Add `pyflysight.log_utils.classify_log_dir` helper for classifying the FlySight hardware rev of a given log directory
 
 ## [v0.2.0]
 ### Added
-* Add parsing pipelines for Flysight V1 track data files
-* #8 Add parsing pipelines for Flysight V2 track & sensor data files
-* #11 Add trimming for Flysight V1 & V2 data files
-* #12 Add simplified CSV log export for Flysight V2 data files
+* Add parsing pipelines for FlySight V1 track data files
+* #8 Add parsing pipelines for FlySight V2 track & sensor data files
+* #11 Add trimming for FlySight V1 & V2 data files
+* #12 Add simplified CSV log export for FlySight V2 data files
 
 ## [v0.1.0]
 Initial release - yay!